pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.13.2",
    "alembic>=1.17.2",
    "bcrypt>=5.0.0",
    "fastapi>=0.123.4",
    "httpx>=0.28.1",
    "litellm>=1.80.7",
    "passlib>=1.7.4",
    "prometheus-client>=0.23.1",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "pyjwt>=2.10.1",
    "python-jose>=3.5.0",
    "python-multipart>=0.0.20",
    "pyyaml>=6.0.3",
    "redis>=7.1.0",
    "sqlalchemy>=2.0.44",
    "structlog>=25.5.0",
    "uvicorn>=0.38.0",
]
